agent_v3/livekit_voice_agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "LiveKit voice AI agent boilerplate with STT-LLM-TTS pipeline"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
]
